doc: Add some docs
