+++ examples/boxes/mod.rs
@@ -1,10 +1,12 @@
 pub use self::bundle::BoxSimulationBundle;
+pub use self::default::{BoxSimulationBundle2, BoxSimulationBundle3};
 pub use self::deletion::BoxDeletionSystem;
 pub use self::emission::EmissionSystem;
 
 mod deletion;
 mod emission;
 mod bundle;
+mod default;
 
 use std::time::{Duration, Instant};
 
@@ -13,6 +15,10 @@ use amethyst::ecs::{Component, DenseVecStorage, VecStorage};
 use amethyst::renderer::{Material, Mesh};
 use rhusics_core::Collider;
 
+/// Collision object type.
+///
+/// Used by collision detection to determine if shapes should be checked for collisions or not.
+/// Only box-box collisions will be processed.
 #[repr(u8)]
 #[derive(Debug, Clone, PartialOrd, PartialEq)]
 pub enum ObjectType {
@@ -36,6 +42,9 @@ impl Component for ObjectType {
     type Storage = VecStorage<Self>;
 }
 
+/// Primitive emitter.
+///
+/// Will emit new primitives based on the internal config.
 pub struct Emitter<P> {
     pub location: P,
     pub last_emit: Instant,
@@ -49,6 +58,7 @@ where
     type Storage = DenseVecStorage<Self>;
 }
 
+/// Internal graphics used for the primitive emission.
 pub struct Graphics {
     pub mesh: Handle<Mesh>,
     pub material: Material,
