+++ src/bundle/mod.rs
@@ -1,4 +1,3 @@
-pub use self::basic_physics::{BasicPhysicsBundle2, BasicPhysicsBundle3,
-                              DefaultBasicPhysicsBundle2, DefaultBasicPhysicsBundle3};
+pub use self::basic_physics::{BasicPhysicsBundle2, BasicPhysicsBundle3};
 
 mod basic_physics;
